+++ draft-packages/stories/ZenNavigationBar.stories.tsx
@@ -1,11 +1,7 @@
 import * as React from "react"
 
 import { Icon } from "@kaizen/component-library"
-import {
-  Link,
-  Menu,
-  ZenNavigationBar,
-} from "@kaizen/component-library/draft/Kaizen/ZenNavigationBar"
+import { Link, Menu, ZenNavigationBar } from "@kaizen/draft-zen-navigation-bar"
 
 const caIcon = require("@kaizen/component-library/icons/ca-monogram.icon.svg")
   .default
@@ -36,6 +32,33 @@ const accountMenuBtn = (
   </div>
 )
 
+const mobileHeaderComponent = (
+  <div
+    style={{
+      width: "100%",
+      padding: "24px 12px",
+      backgroundColor: "#E6F0F7",
+      color: "#0168B3",
+      fontFamily: "sans-serif",
+      boxSizing: "border-box",
+    }}
+  >
+    Switch to performance
+  </div>
+)
+
+const footerComponent = (
+  <div
+    style={{
+      borderTop: "1px solid #eee",
+      height: "inherit",
+      backgroundColor: "#fff",
+    }}
+  >
+    Footer test
+  </div>
+)
+
 export const Default = () => (
   <ZenNavigationBar onNavigationChange={handleNavigationChange}>
     {{
@@ -47,13 +70,14 @@ export const Default = () => (
           href="/"
           badge={{ kind: "new", text: "New" }}
         />,
+      ],
+      secondary: [
         <Link
           text="Inbox"
           href="/"
           badge={{ kind: "notification", text: "5" }}
         />,
       ],
-      secondary: [],
       final: [
         <Link
           icon={supportIcon}
@@ -61,6 +85,7 @@ export const Default = () => (
           href="http://academy.cultureamp.com/"
         />,
         <Link
+          tooltip="Opens in new tab"
           icon={academyIcon}
           text="Academy"
           href="http://academy.cultureamp.com/"
@@ -209,6 +234,127 @@ export const MenuGroup = () => (
                 {
                   label: "Skills",
                   url: "meh",
+                  showArrowIcon: true,
+                },
+              ],
+            },
+          ]}
+        />
+      ),
+    }}
+  >
+    {{
+      primary: [
+        <Link text="Home" href="/" active />,
+        <Link text="Surveys" href="/" />,
+        <Link
+          text="Performance"
+          href="/"
+          badge={{ kind: "new", text: "New" }}
+        />,
+      ],
+      secondary: [
+        <Menu
+          heading="Admin"
+          items={[
+            {
+              label: "Skills",
+              url: "meh",
+            },
+            {
+              label: "Self-reflections",
+              url: "meh",
+            },
+            {
+              title: "Manager requested feedback",
+              items: [
+                {
+                  label: "Request history",
+                  url: "meh",
+                },
+                {
+                  label: "Request feedback",
+                  url: "meh",
+                },
+              ],
+            },
+            {
+              title: "Usage stats",
+              items: [
+                {
+                  label: "Goal stats",
+                  url: "meh",
+                },
+                {
+                  label: "Feedback stats",
+                  url: "meh",
+                },
+              ],
+            },
+          ]}
+        />,
+      ],
+      final: [
+        <Link
+          icon={supportIcon}
+          text="Support"
+          href="http://academy.cultureamp.com/"
+        />,
+        <Link
+          icon={academyIcon}
+          text="Academy"
+          href="http://academy.cultureamp.com/"
+        />,
+        <Menu
+          heading="Custom menu..."
+          items={[
+            {
+              label: "About Culture Amp",
+              url: "https://www.cultureamp.com/",
+            },
+            {
+              label: "Contribute to this guide",
+              url:
+                "https://github.com/cultureamp/cultureamp-style-guide/tree/master/guide",
+            },
+            {
+              label: "Sign out",
+              url: "http://localhost:3000/session/sign_out",
+              method: "delete",
+            },
+            {
+              label: "Stop Masquerading",
+              url: "http://localhost:3000/admin/masquerade/",
+              method: "delete",
+            },
+          ]}
+        >
+          {accountMenuBtn}
+        </Menu>,
+      ],
+    }}
+  </ZenNavigationBar>
+)
+
+export const WithFooterAndHeaderComponents = () => (
+  <ZenNavigationBar
+    onNavigationChange={handleNavigationChange}
+    footerComponent={footerComponent}
+    headerComponent={{
+      mobile: mobileHeaderComponent,
+      desktop: (
+        <Menu
+          icon={caIcon}
+          opaque
+          heading="Engagement"
+          items={[
+            {
+              title: "Switch To",
+              items: [
+                {
+                  label: "Skills",
+                  url: "meh",
+                  showArrowIcon: true,
                 },
               ],
             },