+++ draft-packages/zen-navigation-bar/KaizenDraft/ZenNavigationBar/types.ts
@@ -29,6 +29,8 @@ export type LinkProps = {
   }
   opaque?: boolean
   small?: boolean
+  content?: boolean
+  tooltip?: string
 }
 
 export type MenuProps = {