+++ packages/component-library/draft/Kaizen/ZenNavigationBar/NavigationBar.tsx
@@ -4,13 +4,7 @@ import { ZenControlledOffCanvas } from "@kaizen/component-library/draft/Kaizen/Z
 import classNames from "classnames"
 import Media from "react-media"
 import uuid from "uuid/v4"
-import {
-  LocalBadge,
-  namedBadge,
-  ProductionBadge,
-  StagingBadge,
-  TestBadge,
-} from "./components/Badge"
+import Badge from "./components/Badge"
 import Link from "./components/Link"
 import Menu from "./components/Menu"
 import { LinkClickContext } from "./context"
@@ -21,7 +15,7 @@ const styles = require("./NavigationBar.module.scss")
 type Props = {
   environment?: "production" | "staging" | "test" | "local"
   loading?: boolean
-  colorScheme?: "cultureamp" | "kaizen" | "content"
+  colorScheme?: ColorScheme
   badgeHref?: string
   onNavigationChange: NavigationChange
   headerComponent?: {
@@ -32,6 +26,8 @@ type Props = {
   children?: Navigation
   mobileMaxWidth?: number
 }
+
+export type ColorScheme = "cultureamp" | "kaizen" | "content"
 
 export default class NavigationBar extends React.Component<Props> {
   static displayName = "NavigationBar"
@@ -63,13 +59,13 @@ export default class NavigationBar extends React.Component<Props> {
           {(matches: boolean) =>
             matches ? (
               <ZenControlledOffCanvas
-                headerComponent={
-                  headerComponent ? headerComponent.mobile : this.renderBadge()
-                }
+                headerComponent={this.renderBadge()}
                 footerComponent={this.props.footerComponent}
+                productSwitcher={headerComponent && headerComponent.mobile}
                 links={children}
                 heading="Menu"
                 menuId="menu"
+                colorScheme={colorScheme}
               />
             ) : (
               <header
@@ -130,6 +126,7 @@ export default class NavigationBar extends React.Component<Props> {
         ...linkProps,
         opaque: isFinal,
         small: isFinal,
+        content: this.props.colorScheme === "content",
       },
     }
     const key = "href" in linkProps ? linkProps.href : linkProps.heading
@@ -148,20 +145,10 @@ export default class NavigationBar extends React.Component<Props> {
 
   renderBadge() {
     const {
-      environment = "production",
       loading = false,
       badgeHref = "/",
       colorScheme = "kaizen",
     } = this.props
-
-    const badges = {
-      production: ProductionBadge,
-      staging: StagingBadge,
-      test: TestBadge,
-      local: LocalBadge,
-    }
-
-    const Badge = badges[environment] || namedBadge(environment)
 
     return (
       <Badge loading={loading} href={badgeHref} colorScheme={colorScheme} />