+++ packages/component-library/stories/ElmTable.stories.tsx
@@ -1,0 +1,5 @@
+import { loadElmStories } from "@cultureamp/elm-storybook"
+
+loadElmStories("Table (Elm)", module, require("./Table.stories.elm"), [
+  "Default",
+])