+++ draft-packages/dropdown-menu/CHANGELOG.md
@@ -3,6 +3,100 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.2.4...@kaizen/draft-dropdown-menu@1.2.5) (2020-05-07)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.2.3...@kaizen/draft-dropdown-menu@1.2.4) (2020-05-06)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.2.2...@kaizen/draft-dropdown-menu@1.2.3) (2020-05-06)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.2.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.2.1...@kaizen/draft-dropdown-menu@1.2.2) (2020-05-05)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.2.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.2.0...@kaizen/draft-dropdown-menu@1.2.1) (2020-05-05)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+# [1.2.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.1.6...@kaizen/draft-dropdown-menu@1.2.0) (2020-05-05)
+
+
+### Features
+
+* Updated all draft packages to accept all minor releases ([#451](https://github.com/cultureamp/kaizen-design-system/issues/451)) ([80453f6](https://github.com/cultureamp/kaizen-design-system/commit/80453f6c04300dcef61c14e39200ce154863eb0d))
+
+
+
+
+
+## [1.1.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.1.5...@kaizen/draft-dropdown-menu@1.1.6) (2020-05-04)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.1.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.1.4...@kaizen/draft-dropdown-menu@1.1.5) (2020-05-04)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.1.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.1.3...@kaizen/draft-dropdown-menu@1.1.4) (2020-05-04)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.1.2...@kaizen/draft-dropdown-menu@1.1.3) (2020-05-04)
+
+**Note:** Version bump only for package @kaizen/draft-dropdown-menu
+
+
+
+
+
+## [1.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.1.1...@kaizen/draft-dropdown-menu@1.1.2) (2020-05-01)
+
+
+### Bug Fixes
+
+* draft-package elm imports ([#444](https://github.com/cultureamp/kaizen-design-system/issues/444)) ([16c9ffd](https://github.com/cultureamp/kaizen-design-system/commit/16c9ffde3ecd6d4f361fb2ab40bf4452a77b0e8a))
+
+
+
+
+
 ## [1.1.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-dropdown-menu@1.1.0...@kaizen/draft-dropdown-menu@1.1.1) (2020-05-01)